requirements.txt
asgiref==3.11.1
Pillow>=10.0.0
boto3==1.42.75
botocore==1.42.75
Django==4.2.29
django-anymail==10.3
djangorestframework==3.16.1
dotenv==0.9.9
jmespath==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==1.26.20
